Disable gh-pages pushes, since it's spewing lots of errors.
It looks like all our frequent pushes amount to 1.5GB of git data, and
gh-pages is triggering build failures. We could push to a non gh-pages
branch, but at this point, I think git is probably not the right place
to be storing these hyper-frequent updates. The Postgres auditing should
hopefully provide similar details if we ever want to dig into historical
data.