Refresh walmart stores metadata on a regular basis.
Rework things so that Walmart stores metadata is regularly updated (by
searching for new stores) so that we catch new stores that begin
carrying the vaccine.

Also rework things a bit to make the appointment refresh process more
standardized and fix some authentication issues to try and make this
more resilient to proxying issues.