Look into the future for 2nd dose appointments at Walgreens.
This is to fix issues with 1st doses not actually being bookable, since
2nd doses are not available. See
https://github.com/GUI/covid-vaccine-spotter/issues/61

This should solve the issue, but I still need to think about how the UI
should work for this. This currently hides the appointments that don't
have 2nd doses as not having any appointments. But I think we may want
some way to filter in the UI to see these 2nd-dose only locations for
people that may just need a 2nd dose.