Set Albertsons last updated time to last fetched time.
I thought maybe it would be better to reflect Albertson's last-modified
date, but that appears to have gone un-updated for a while, so now I'm
setting this back to the last fetched date so it doesn't appear like
we're being blocked. Still not quite sure what the right thing here is.