Re-enable gzip compression before uploading to bucket.
To cut down on bandwidth. I think the issues causing the revert in
4b50911aa453ab43f9b7f5e07b9f2528617a6127, were more due to the order of
the rollout, with some of the files not being properly re-uploaded with
the proper "Content-Encoding: gzip" header. I think I've manually forced
everything to be reuploaded during this rollout (by temporary rolling
things out with rclone's "--no-check-dest" flag), so I think this is now
working properly.