Fix Health Mart and PriceChopper failing due to invalid SSL cert.
The scrcxp.pdhi.com domain's SSL certificate that was updated today is
missing an intermediate SSL cert (that browsers will download, but curl
does not). Install that intermediate in the root certificate store to
get things to be trusted again.

As part of this, abstract out some of the common curl settings so it's
easier to set custom CA paths.