Fix missing Albertsons store details for new stores.
Since revamping the Albertsons process to be based on the new file data,
if new stores presented themselves, we no longer had any logic necessary
to parse the overall store details (we were reliant on the store details
already being present in the database). This should fix that so that we
now parse any new store location details out of the new data feed.