Add full appointment details to map popup.
With the shift towards paginating the results, it's trickier for the
popup to have a link to the details in the right column (since we'd have
to determine what page it's on). So move towards just showing identical
content in the map popup, which should also hopefully make things more
directly usable and obvious anyway.