Fix for writing US-wide API JSON file again.
In bdce1224e3770a3f3b3fd036efaa7c5beac8e54b, we disabled writing the
US.json file, since we had exceeding PostgreSQL's memory limits for
outputting everything directly from the database.

This changes things around so that we still build the high-level
metadata with the same queries, but don't output the individual
features/stores in the same output. Instead, we stream back each store
individually, and append it to the appropriate JSON files (both US-wide
and individual states) using streaming writes, so we never have to read
all of this data fully into memory. This also wraps things in a
transaction, so there should now be better consistency between the US
file and the state files.