Fix ThriftyWhite breaking due to new non-store location.
They have a headquarters location that seems to be offering the vaccine,
but isn't marked as a store, so it wasn't in our previous stores
database. So handle this by creating new stores as needed based on the
appointment data.

Also refresh their stores on a regular basis, since it's quick and easy.