Attempt to get Walgreens working again. Sigh.
They no longer require logins, so try to use the non logged in session.

But it seems like there's more rate limiting going on perhaps, so while
we will still parallelize the processing on our end, try to limit the
concurrency of the actual requests. We'll see how slow this ends up
being or if this really ends up working.